+++ CHANGELOG.rst
@@ -18,3 +18,8 @@
 1.4 (2015-09-16)
 ================
 * fix error in ad request
+
+1.5.0 (2015-09-22)
+==================
+* rename AFNetworking
+* fix impression tracking